requirements.txt
load_dotenv
requests
groq
pydub